+++ src/components/Pagination/PageInfo.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { UUI } from '../../core/uui';
-import { PaginationContext } from './Pagination';
+import { PaginationContext } from './PaginationContext';
 
 export interface PageInfoFeatureProps {
   /**
